added run method used Values in waterfall steps
