+++ samples/csharp_dotnetcore/05.multi-turn-prompt/DialogExtensions.cs
@@ -1,0 +1,27 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Schema;
+
+namespace Microsoft.BotBuilderSamples
+{
+    public static class DialogExtensions
+    {
+        public static async Task Run(this Dialog dialog, ITurnContext<IMessageActivity> turnContext, ConversationState conversationState, CancellationToken cancellationToken)
+        {
+            var dialogSet = new DialogSet(conversationState.CreateProperty<DialogState>("DialogState"));
+            dialogSet.Add(dialog);
+
+            var dialogContext = await dialogSet.CreateContextAsync(turnContext, cancellationToken);
+            var results = await dialogContext.ContinueDialogAsync(cancellationToken);
+            if (results.Status == DialogTurnStatus.Empty)
+            {
+                await dialogContext.BeginDialogAsync(dialog.Id, null, cancellationToken);
+            }
+        }
+    }
+}
