--- samples/csharp_dotnetcore/05.multi-turn-prompt/Dialogs/RootDialog.cs
@@ -1,167 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Builder.Dialogs;
-
-namespace Microsoft.BotBuilderSamples
-{
-    public class RootDialog : ComponentDialog
-    {
-        public RootDialog(UserState userState)
-            : base("root")
-        {
-            UserProfileAccessor = userState.CreateProperty<UserProfile>("UserProfile");
-
-            // This array defines how the Waterfall will execute.
-            var waterfallSteps = new WaterfallStep[]
-            {
-                NameStepAsync,
-                NameConfirmStepAsync,
-                AgeStepAsync,
-                ConfirmStepAsync,
-                SummaryStepAsync,
-            };
-
-            // Add named dialogs to the DialogSet. These names are saved in the dialog state.
-            AddDialog(new WaterfallDialog("details", waterfallSteps));
-            AddDialog(new TextPrompt("name"));
-            AddDialog(new NumberPrompt<int>("age", AgePromptValidatorAsync));
-            AddDialog(new ConfirmPrompt("confirm"));
-        }
-
-        public IStatePropertyAccessor<UserProfile> UserProfileAccessor { get; private set; }
-
-        /// <summary>
-        /// One of the functions that make up the <see cref="WaterfallDialog"/>.
-        /// </summary>
-        /// <param name="stepContext">The <see cref="WaterfallStepContext"/> gives access to the executing dialog runtime.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/>.</param>
-        /// <returns>A <see cref="DialogTurnResult"/> to communicate some flow control back to the containing WaterfallDialog.</returns>
-        private static async Task<DialogTurnResult> NameStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            // WaterfallStep always finishes with the end of the Waterfall or with another dialog; here it is a Prompt Dialog.
-            // Running a prompt here means the next WaterfallStep will be run when the users response is received.
-            return await stepContext.PromptAsync("name", new PromptOptions { Prompt = MessageFactory.Text("Please enter your name.") }, cancellationToken);
-        }
-
-        /// <summary>
-        /// One of the functions that make up the <see cref="WaterfallDialog"/>.
-        /// </summary>
-        /// <param name="stepContext">The <see cref="WaterfallStepContext"/> gives access to the executing dialog runtime.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/>.</param>
-        /// <returns>A <see cref="DialogTurnResult"/> to communicate some flow control back to the containing WaterfallDialog.</returns>
-        private async Task<DialogTurnResult> NameConfirmStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            // Get the current profile object from user state.
-            var userProfile = await UserProfileAccessor.GetAsync(stepContext.Context, () => new UserProfile(), cancellationToken);
-
-            // Update the profile.
-            userProfile.Name = (string)stepContext.Result;
-
-            // We can send messages to the user at any point in the WaterfallStep.
-            await stepContext.Context.SendActivityAsync(MessageFactory.Text($"Thanks {stepContext.Result}."), cancellationToken);
-
-            // WaterfallStep always finishes with the end of the Waterfall or with another dialog; here it is a Prompt Dialog.
-            return await stepContext.PromptAsync("confirm", new PromptOptions { Prompt = MessageFactory.Text("Would you like to give your age?") }, cancellationToken);
-        }
-
-        /// <summary>
-        /// One of the functions that make up the <see cref="WaterfallDialog"/>.
-        /// </summary>
-        /// <param name="stepContext">The <see cref="WaterfallStepContext"/> gives access to the executing dialog runtime.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/>.</param>
-        /// <returns>A <see cref="DialogTurnResult"/> to communicate some flow control back to the containing WaterfallDialog.</returns>
-        private async Task<DialogTurnResult> AgeStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            if ((bool)stepContext.Result)
-            {
-                // User said "yes" so we will be prompting for the age.
-                // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is a Prompt Dialog.
-                return await stepContext.PromptAsync(
-                    "age",
-                    new PromptOptions
-                    {
-                        Prompt = MessageFactory.Text("Please enter your age."),
-                        RetryPrompt = MessageFactory.Text("The value entered must be less than 150."),
-                    },
-                    cancellationToken);
-            }
-            else
-            {
-                // User said "no" so we will skip the next step. Give -1 as the age.
-                return await stepContext.NextAsync(-1, cancellationToken);
-            }
-        }
-
-        /// <summary>
-        /// One of the functions that make up the <see cref="WaterfallDialog"/>.
-        /// </summary>
-        /// <param name="stepContext">The <see cref="WaterfallStepContext"/> gives access to the executing dialog runtime.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/>.</param>
-        /// <returns>A <see cref="DialogTurnResult"/> to communicate some flow control back to the containing WaterfallDialog.</returns>
-        private async Task<DialogTurnResult> ConfirmStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            // Get the current profile object from user state.
-            var userProfile = await UserProfileAccessor.GetAsync(stepContext.Context, () => new UserProfile(), cancellationToken);
-
-            // Update the profile.
-            userProfile.Age = (int)stepContext.Result;
-
-            // We can send messages to the user at any point in the WaterfallStep.
-            if (userProfile.Age == -1)
-            {
-                await stepContext.Context.SendActivityAsync(MessageFactory.Text($"No age given."), cancellationToken);
-            }
-            else
-            {
-                // We can send messages to the user at any point in the WaterfallStep.
-                await stepContext.Context.SendActivityAsync(MessageFactory.Text($"I have your age as {userProfile.Age}."), cancellationToken);
-            }
-
-            // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is a Prompt Dialog.
-            return await stepContext.PromptAsync("confirm", new PromptOptions { Prompt = MessageFactory.Text("Is this ok?") }, cancellationToken);
-        }
-
-        /// <summary>
-        /// One of the functions that make up the <see cref="WaterfallDialog"/>.
-        /// </summary>
-        /// <param name="stepContext">The <see cref="WaterfallStepContext"/> gives access to the executing dialog runtime.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/>.</param>
-        /// <returns>A <see cref="DialogTurnResult"/> to communicate some flow control back to the containing WaterfallDialog.</returns>
-        private async Task<DialogTurnResult> SummaryStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
-        {
-            if ((bool)stepContext.Result)
-            {
-                // Get the current profile object from user state.
-                var userProfile = await UserProfileAccessor.GetAsync(stepContext.Context, () => new UserProfile(), cancellationToken);
-
-                // We can send messages to the user at any point in the WaterfallStep.
-                if (userProfile.Age == -1)
-                {
-                    await stepContext.Context.SendActivityAsync(MessageFactory.Text($"I have your name as {userProfile.Name}."), cancellationToken);
-                }
-                else
-                {
-                    await stepContext.Context.SendActivityAsync(MessageFactory.Text($"I have your name as {userProfile.Name} and age as {userProfile.Age}."), cancellationToken);
-                }
-            }
-            else
-            {
-                // We can send messages to the user at any point in the WaterfallStep.
-                await stepContext.Context.SendActivityAsync(MessageFactory.Text("Thanks. Your profile will not be kept."), cancellationToken);
-            }
-
-            // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is the end.
-            return await stepContext.EndDialogAsync(cancellationToken: cancellationToken);
-        }
-
-        private static Task<bool> AgePromptValidatorAsync(PromptValidatorContext<int> promptContext, CancellationToken cancellationToken)
-        {
-            // This condition is our validation rule. You can also change the value at this point.
-            return Task.FromResult(promptContext.Recognized.Value < 150);
-        }
-    }
-}
