+++ samples/csharp_dotnetcore/05.multi-turn-prompt/Dialogs/UserProfileDialog.cs
@@ -1,0 +1,124 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
+
+namespace Microsoft.BotBuilderSamples
+{
+    public class UserProfileDialog : ComponentDialog
+    {
+        private IStatePropertyAccessor<UserProfile> _userProfileAccessor;
+
+        public UserProfileDialog(UserState userState)
+            : base("root")
+        {
+            _userProfileAccessor = userState.CreateProperty<UserProfile>("UserProfile");
+
+            // This array defines how the Waterfall will execute.
+            var waterfallSteps = new WaterfallStep[]
+            {
+                NameStepAsync,
+                NameConfirmStepAsync,
+                AgeStepAsync,
+                ConfirmStepAsync,
+                SummaryStepAsync,
+            };
+
+            // Add named dialogs to the DialogSet. These names are saved in the dialog state.
+            AddDialog(new WaterfallDialog("details", waterfallSteps));
+            AddDialog(new TextPrompt("name"));
+            AddDialog(new NumberPrompt<int>("age", AgePromptValidatorAsync));
+            AddDialog(new ConfirmPrompt("confirm"));
+        }
+
+        private static async Task<DialogTurnResult> NameStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        {
+            // WaterfallStep always finishes with the end of the Waterfall or with another dialog; here it is a Prompt Dialog.
+            // Running a prompt here means the next WaterfallStep will be run when the users response is received.
+            return await stepContext.PromptAsync("name", new PromptOptions { Prompt = MessageFactory.Text("Please enter your name.") }, cancellationToken);
+        }
+
+        private async Task<DialogTurnResult> NameConfirmStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        {
+            stepContext.Values["name"] = (string)stepContext.Result;
+
+            // We can send messages to the user at any point in the WaterfallStep.
+            await stepContext.Context.SendActivityAsync(MessageFactory.Text($"Thanks {stepContext.Result}."), cancellationToken);
+
+            // WaterfallStep always finishes with the end of the Waterfall or with another dialog; here it is a Prompt Dialog.
+            return await stepContext.PromptAsync("confirm", new PromptOptions { Prompt = MessageFactory.Text("Would you like to give your age?") }, cancellationToken);
+        }
+
+        private async Task<DialogTurnResult> AgeStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        {
+            if ((bool)stepContext.Result)
+            {
+                // User said "yes" so we will be prompting for the age.
+                // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is a Prompt Dialog.
+                var promptOptions = new PromptOptions
+                {
+                    Prompt = MessageFactory.Text("Please enter your age."),
+                    RetryPrompt = MessageFactory.Text("The value entered must be greater than 0 and less than 150."),
+                };
+
+                return await stepContext.PromptAsync("age", promptOptions, cancellationToken);
+            }
+            else
+            {
+                // User said "no" so we will skip the next step. Give -1 as the age.
+                return await stepContext.NextAsync(-1, cancellationToken);
+            }
+        }
+
+        private async Task<DialogTurnResult> ConfirmStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        {
+            stepContext.Values["age"] = (int)stepContext.Result;
+
+            // We can send messages to the user at any point in the WaterfallStep.
+            //var msg = userProfile.Age == -1 ? "No age given." : $"I have your age as {userProfile.Age}.";
+            var msg = (int)stepContext.Values["age"] == -1 ? "No age given." : $"I have your age as {stepContext.Values["age"]}.";
+
+            // We can send messages to the user at any point in the WaterfallStep.
+            await stepContext.Context.SendActivityAsync(MessageFactory.Text(msg), cancellationToken);
+
+            // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is a Prompt Dialog.
+            return await stepContext.PromptAsync("confirm", new PromptOptions { Prompt = MessageFactory.Text("Is this ok?") }, cancellationToken);
+        }
+
+        private async Task<DialogTurnResult> SummaryStepAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
+        {
+            if ((bool)stepContext.Result)
+            {
+                // Get the current profile object from user state.
+                var userProfile = await _userProfileAccessor.GetAsync(stepContext.Context, () => new UserProfile(), cancellationToken);
+
+                userProfile.Name = (string)stepContext.Values["name"];
+                userProfile.Age = (int)stepContext.Values["age"];
+
+                var msg = $"I have your name as {userProfile.Name}.";
+                if (userProfile.Age != -1)
+                {
+                    msg += " And age as ${userProfile.Age}.";
+                }
+
+                await stepContext.Context.SendActivityAsync(MessageFactory.Text(msg), cancellationToken);
+            }
+            else
+            {
+                await stepContext.Context.SendActivityAsync(MessageFactory.Text("Thanks. Your profile will not be kept."), cancellationToken);
+            }
+
+            // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is the end.
+            return await stepContext.EndDialogAsync(cancellationToken: cancellationToken);
+        }
+
+        private static Task<bool> AgePromptValidatorAsync(PromptValidatorContext<int> promptContext, CancellationToken cancellationToken)
+        {
+            // This condition is our validation rule. You can also change the value at this point.
+            return Task.FromResult(promptContext.Recognized.Value >= 0 && promptContext.Recognized.Value < 150);
+        }
+    }
+}
